feat(debug): trace every logic random draw, from a release build
A lockstep mismatch was localised to the LIVE logic-RNG state CRC -- the two
peers had drawn a different NUMBER of logic random values by frame 4300, with
every object, the partition manager, the player list and every player's money
still byte-identical. That says "one peer took a draw the other did not" and
says nothing about WHICH draw, and reading 243 call sites did not find it.

GENERALS_RANDOM_TRACE=<path> now writes one line per logic draw:

    <frame> i <value> <lo> <hi> <file>:<line>     integer
    <frame> r <float-bits-hex> <file>:<line>      real

Diff two peers' files and the first differing line names the call site. Reals
are logged as raw bits rather than %f, because a decimal rendering can hide a
low-mantissa difference, which is exactly the kind worth seeing. Paths are
reduced to a basename so two machines with different source roots -- or
different architectures -- do not differ on every line. The Unchanged variants
are deliberately not traced: they roll a copy of the seed and cannot move the
CRC.

WHY NOT THE EXISTING DEBUG_RANDOM_LOGIC

It logs through DEBUG_LOG, which needs RTS_DEBUG_LOGGING=ON, and that
configuration DOES NOT COMPILE on Linux. Enabling it pulls in a chain of
debug-only code that only 32-bit MSVC has ever built. Three of those are fixed
here because they are small and unambiguous:

  * wwdebug.cpp used strlen with no <string.h>; the MSVC headers supplied it
    transitively and no release build compiles those asserts.
  * sphereobj.cpp asserted pointer equality via (int) casts, a 64-bit
    truncation. Now compares the pointers directly.
  * BOGUSPTR cast a pointer to unsigned int. Now uintptr_t.

The chain does not end there -- GameClient.cpp reads Windows MEMORYSTATUS
fields that do not exist here -- so fixing the debug configuration is left as
its own job. The trace deliberately does not depend on it: one FILE*, plain
fprintf, no subsystem. It therefore runs in the RELEASE build every peer
already has, at release speed, with no second binary per machine. Unset, the
cost is one pointer test per draw.

The four hand-edited toggles in RandomValue.cpp are also now settable from the
build (RTS_DEBUG_RANDOM_LOGIC and friends, plus a linux64-randomtrace preset
that switches off the profile/stacktrace options whose code does not compile).
Previously an unconditional #undef block discarded anything defined on the
command line, so -DDEBUG_RANDOM_LOGIC silently did nothing.

Verified: release build links, a 600-frame headless match traced 61 draws
across 8 call sites (AIStates.cpp:1341, Locomotor.cpp:674-676,
PartitionManager.cpp:3986, StealthDetectorUpdate.cpp:90, AIUpdate.cpp:4608,
AIPlayer.cpp:2991), and an engine sitting in a lobby without starting traces
nothing, which is correct -- no logic frames, no draws.

+++ Core/GameEngine/Include/Common/GameCommon.h
@@ -306,7 +306,12 @@ inline VeterancyLevelFlags clearVeterancyLevelFlag(VeterancyLevelFlags flags, Ve
 }
 
 // ----------------------------------------------------------------------------------------------
-#define BOGUSPTR(p) ((((unsigned int)(p)) & 1) != 0)
+// Tests the low bit, which the DLINK debug checks use to mark a deliberately
+// bogus pointer. uintptr_t, not unsigned int: the original truncates a 64-bit
+// pointer, which is an error rather than a warning on GCC/Clang. It only bites
+// in a build with debug checks enabled, which is why a release build never
+// noticed.
+#define BOGUSPTR(p) ((((uintptr_t)(p)) & 1) != 0)
 
 // ----------------------------------------------------------------------------------------------
 #define MAKE_DLINK_HEAD(OBJCLASS, LISTNAME)																						\
